GEOMESA-381 Localize TServer classloading trick
* The classloader trick which needs to happen on the tablet servers is broken out.
It should not be run by the GeoMesa GeoServer plugin.
