+++ geomesa-core/src/main/scala/org/locationtech/geomesa/core/iterators/TServerClassLoader.scala
@@ -1,0 +1,36 @@
+package org.locationtech.geomesa.core.iterators
+
+import com.typesafe.scalalogging.slf4j.Logger
+import org.apache.commons.vfs2.impl.VFSClassLoader
+import org.geotools.factory.GeoTools
+
+object TServerClassLoader {
+
+  val initialized = new ThreadLocal[Boolean] {
+    override def initialValue(): Boolean = false
+  }
+
+  def initClassLoader(log: Logger) =
+    if(!initialized.get()) {
+      try {
+        log.trace("Initializing classLoader")
+        // locate the geomesa-distributed-runtime jar
+        val cl = this.getClass.getClassLoader
+        cl match {
+          case vfsCl: VFSClassLoader =>
+            val url = vfsCl.getFileObjects.map(_.getURL).filter {
+              _.toString.contains("geomesa-distributed-runtime")
+            }.head
+            if (log != null) log.debug(s"Found geomesa-distributed-runtime at $url")
+            val u = java.net.URLClassLoader.newInstance(Array(url), vfsCl)
+            GeoTools.addClassLoader(u)
+          case _ =>
+        }
+      } catch {
+        case t: Throwable =>
+          if(log != null) log.error("Failed to initialize GeoTools' ClassLoader ", t)
+      } finally {
+        initialized.set(true)
+      }
+    }
+}
